backend/accounts/serializers.py
from rest_framework import serializers
from .models import UserProfile

class ProfilePictureSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = ['profile_picture']