requirements.txt
# Core Django Setup
Django>4.2,<5.0
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
drf-yasg
Pillow


